One more fix for Soyes XS phones (but see disclaimer below)
The bluetooth implementation of this phone calls the gatt callbacks
*before* the connect call returns to the client.  This is incorrect because
the client won't have a reference to the gatt at that time.

Fortunately, we only need to check gatt on disconnect, so I moved that later.

But one problem I've noticed in my testing.  Sometimes this phone stops
being able to scan for BLE devices.  The only fix I've found is to click to
turn bluetooth off briefly and then back on.

A major problem with this phone is that if you reboot the phone it seems
the time the phone has forgotten its paring data and devices must
be re-paired.  This is a huge bummer and I don't think my app can fix this.