Display PKI encryption status on messages
This commit adds an icon to messages to indicate whether they are encrypted with PKI.

The icon is a lock if
 the message is encrypted and an unlocked lock if it is not.
The icon is displayed next to the message status icon.
A content description is also added to the icon to provide more information to users who use screen readers.