Change max message length to 200
IPhone app uses 200 chars as it's message limit.

This was last updated after the introduction of PKI. Testing there revealed that with PKI the real max now is about 225, but reliability starts to dip before then. I confirmed this in testing wirh Android on a few devices just now.

Change Android app to match iPhone.