Refactor: Use string resources for config and module titles
The titles for config and module settings are now retrieved from string resources instead of being hardcoded.
This commit adds new string resources for the titles of config and module routes, such as "User", "Channels", "Device", "MQTT", and "Serial".
It also changes `ModuleRoute` and `ConfigRoute` classes to use `stringRes` to refer to these string resources, and changes `RadioConfig` class to use `stringResource` to retrieve the title.