refactor: Remove BaseScaffold and improve text input
The BaseScaffold composable has been removed as it is no longer needed. The functionality provided by BaseScaffold is now handled directly within the screens that used it, specifically the Message screen, which now uses the standard `Scaffold`.

Additionally, the text input component in the message screen has been refactored to use `OutlinedTextField` with a trailing icon for the send button. This provides a more standard and consistent UI element for text input. The clipboard handling has also been updated to use the new ClipEntry API.

The Message screen title is now handled by the UIViewModel's setTitle method.

The title display logic in the main navigation graph has been adjusted to show title for the Messages route.

Clipboard handling in the Channel and LinkedCoordinates components has also been updated to use the new ClipEntry API.