Enable 'Unmessageable' toggle based on firmware version
This commit updates the logic for enabling the 'Unmessageable' toggle in the user configuration and node item views.

Previously, the toggle was enabled if the `isUnmessagable` field was present in the user configuration. Now, it is enabled if the device firmware version is 2.6.8 or newer.

This change ensures that the 'Unmessageable' feature is only available on devices that support it.