Support new `is_unmessagable` flag for device roles
This commit updates the `NodeItem` composable to correctly determine if a node is unmessageable.

Previously, it checked for firmware version "2.6.8" and the `isUnmessagable` user property.

Now, it prioritizes the new `hasIsUnmessagable()` flag on the user object. If this flag is not present, it falls back to checking the `isUnmessageableRole()` and the firmware version, defaulting to unmessageable if the firmware version is unavailable.