Refactor: Use FLAG_ACTIVITY_SINGLE_TOP for MainActivity intents
This change ensures that when `MainActivity` is launched via an intent (e.g., from a notification or shortcut), an existing instance of the activity is brought to the front instead of creating a new one. This improves the user experience by preventing multiple instances of the main activity from being created.