Enforce valid names for UserConfig and remove `enabled` prop from PreferenceFooter `Cancel`
The `PreferenceFooter` component now relies on the `enabled` state of the `UserConfigItemList` to determine if the "Save" button should be active. This change ensures that the save button is only enabled when both the long and short names are valid and there are changes to save.

Additionally, the redundant `enabled` property was removed from the "Cancel" button within `PreferenceFooter` as we want to allow the user to be able to cancel changes even when invalid.

Signed-off-by: James Rich <2199651+jamesarich@users.noreply.github.com>