refactor: ClientNotification messages as an alert dialog
RadioConfig error and therefore ClientNotification messages were previously shown as a temporary snackbar, but now they are shown as a dialog.

The dialog is not dismissable, so the user has to acknowledge the message before they can continue.

This provides a more prominent exposure and acknowledgement of Client Notification messages that may require a user's action or attention.

A new string `client_notification` has been added for the title of the alert.

Signed-off-by: James Rich <2199651+jamesarich@users.noreply.github.com>