fix (#2128): MQTT map reporting interval input
This commit fixes the MQTT map reporting interval input. The interval is now displayed in the UI and can be changed by the user.

A validation check has been added to ensure the interval is not less than 3600 seconds (1 hour). If an invalid value is entered, the field displays the error state.

This change also updates the consent check for map reporting to consider the publish interval. If the interval is less than 1 hour, consent will be considered invalid.

Signed-off-by: James Rich <2199651+jamesarich@users.noreply.github.com>