feat: Add scroll to bottom button in message list
This commit introduces a "scroll to bottom" floating action button in the message list view.

The button appears when the user can scroll backward (i.e., when not at the very bottom of the list).
Clicking the button smoothly animates the list to the newest message.

The `LazyListState` is now initialized in the `MessageScreen` and passed down to `MessageList` to enable this functionality.
A new string resource `scroll_to_bottom` has also been added.

Signed-off-by: James Rich <2199651+jamesarich@users.noreply.github.com>