docs: update mesh_service_example README
The README for `mesh_service_example` has been updated to clarify that it's an example of an app using the AIDL Mesh Service provided by Meshtastic-Android, rather than an implementation of the service itself.

It now also links to the AIDL definition in the main app module and removes a reference to best practices in module development, which is less relevant to an example consumer app.

Signed-off-by: James Rich <2199651+jamesarich@users.noreply.github.com>