Refactor: Update Gradle cache keys and task execution in CI
This commit updates the Gradle cache keys in the reusable Android build and test workflows to be more specific and include `libs.versions.toml` and `.kts` files.

It also modifies the execution of Gradle tasks in the build workflow to explicitly target the `:app` module for `detekt`, `lint`, `assemble`, and `test` tasks.

Signed-off-by: James Rich <2199651+jamesarich@users.noreply.github.com>