refactor(Main): Simplify top-level destination check
Refactor the `isTopLevel()` extension function on `NavDestination` to directly check against a predefined list of top-level `Route` objects instead of iterating through `TopLevelDestination.entries`.

This change also updates the `TopAppBar` composable to use the revised `isTopLevel()` check and ensures that menu actions are only displayed when `currentDestination` is not null.

Signed-off-by: James Rich <2199651+jamesarich@users.noreply.github.com>