feat(#2488): Add 🔔 quick chat to database so it can be deleted
This commit introduces a database migration to add a predefined "alert" quick chat message (🔔) to the `quick_chat` table if it doesn't already exist.

The hardcoded alert action has been removed from the `Message.kt` UI, and the quick chat actions are now solely sourced from the database.

This allows the user to delete the 🔔 if they want.

Signed-off-by: James Rich <2199651+jamesarich@users.noreply.github.com>