fix(radioconfig): request location permission when setting node position
When setting the node position to the phone's current location, the app now checks for and requests the `ACCESS_FINE_LOCATION` permission if it has not already been granted. Previously, the button to set the location would be disabled if the location was unknown, which could occur if the permission was missing.

Signed-off-by: James Rich <2199651+jamesarich@users.noreply.github.com>