refactor(radioconfig): Change locationFlow to suspend function
The `locationFlow` in `RadioConfigViewModel` has been changed from a `StateFlow` to a suspend function `getCurrentLocation()`. This change simplifies the logic for obtaining the current location and ensures that location updates are only requested when needed.

The `PositionConfigScreen` has been updated to use the new `getCurrentLocation()` function and to handle location permission requests.
The `PositionConfigItemList` has been updated to accept a suspend function `onUseCurrentLocation` lambda.

Signed-off-by: James Rich <2199651+jamesarich@users.noreply.github.com>