refactor(db): update NodeDB persistence logic
This commit refactors the persistence logic for the NodeDB.
The `installNodeDB` function in `RadioConfigRepository` and `NodeRepository` is replaced with `installMyNodeEntity`.
This change ensures that `MyNodeEntity` is installed separately from the rest of the `NodeEntity` list.
Additionally, a call to `radioConfigRepository.installMyNodeEntity(myNodeInfo!!)` is added in `MeshService` after `myNodeInfo` is updated.
A comment regarding `hasNodeDB` has been removed as it is no longer relevant with the new persistence logic.

Signed-off-by: James Rich <2199651+jamesarich@users.noreply.github.com>