fix(service): ensure onConnected is called after myNodeInfo is set
The `onConnected` call was moved from `onConfigComplete` to `setMyNodeInfo`
to guarantee that `myNodeInfo` is populated before `onConnected` is invoked.
This prevents potential race conditions or null pointer exceptions that could occur
if `onConnected` relies on `myNodeInfo` being available.

The redundant `onConnected` call in `sendNodeInfoOnlyRequest` was also removed
as it is now handled correctly within `setMyNodeInfo`.

Signed-off-by: James Rich <2199651+jamesarich@users.noreply.github.com>