feat(logging): Remove redundant Report Bug button and related functionality
This commit removes the "Report Bug" button from the connections screen and the associated `reportError` function from the `Logging` class. The functionality to report bugs through a dialog and send reports to a crash reporter service has been removed.

The native exception capturing offered by Crashlytics and Datadog are sufficient.

Signed-off-by: James Rich <2199651+jamesarich@users.noreply.github.com>